feat: written skeleton for business logic tests
