+++ app/service/storage/storage.py
@@ -5,11 +5,11 @@ from app.entities.storage import Storage
 from app.errors.service_error.storage_error import StockClosed, ImpossibleCloseStock
 from app.errors.service_error.tool_error import ToolBroken
 from app.errors.service_error.construction_error import ConstructionClosed, ResponsibleAbsent
-from database.crud.constructionCRUD import ConstructionCRUD
-from database.crud.toolCRUD import ToolCRUD
-from database.crud.storageCRUD import StorageCRUD
-from database.crud.constructionCRUD import ConstructionCRUD
-from database.crud.toolCRUD import ToolCRUD
+from app.database.crud.constructionCRUD import ConstructionCRUD
+from app.database.crud.toolCRUD import ToolCRUD
+from app.database.crud.storageCRUD import StorageCRUD
+from app.database.crud.constructionCRUD import ConstructionCRUD
+from app.database.crud.toolCRUD import ToolCRUD
 
 
 class StorageStatus(enum.Enum):
