+++ app/service/tool/tool.py
@@ -5,8 +5,8 @@ from app.entities.storage import Storage
 from app.errors.service_error.tool_error import ToolBroken
 from app.errors.service_error.storage_error import StockClosed
 from app.errors.service_error.construction_error import ConstructionClosed, ResponsibleAbsent
-from database.crud.constructionCRUD import ConstructionCRUD
-from database.crud.toolCRUD import ToolCRUD
+from app.database.crud.constructionCRUD import ConstructionCRUD
+from app.database.crud.toolCRUD import ToolCRUD
 
 
 class ToolStatus(enum.Enum):
