+++ tests.py
@@ -1,2 +1,2 @@
-# from app.tests.service_tests.test_essence import *
-from app.tests.database_test.test_database import *
+from app.tests.service_tests.test_logic import TestBusinessLogic
+from app.tests.database_test.test_database import TestDatabase
